Update Ramiro Merriweather-Italic LC and Artur caps.glyphs
I made glyphs interpolate again that had stapped ( The original outlines are in the mask) and I tweaked spacing. The really changed glyphs are in dark blue. The ones I checked but didn't alter are in light blue. The biggest change was to the z which I decided felt better when it was quite a bit narrower.