Update Ramiro Merriweather-Italic LC and Artur caps.glyphs
Black Display - Adjusted the inside of o to match overall slant and the spacing of t. made the left side of a darker for more even color. Fixed the flow at the bottom of t and apparent angle of t. Dropped the left side of c. Added ezset. e- adusted upper right of the eye for better unity. and reduced left side bearing.