Update Ramiro Merriweather-Italic LC and Artur caps.glyphs
Light Display - adjusted a to match d better. adjusted the stem of t. Adjusted many stems a b d h i k j to be 140 units wide. raised the center join or crotch of k. Reduced the stroke weight in the middle of s middle. reduced space to a width of 600 units.