Merriweather-Italic.Cyrillic SC glyphs - Cyrillic lc, UC and SC
This file has changes to Cyrillic lc, UC and SC. Everything colored dark grey is new. The things that are new in latin mostly concern anchor placement or diacritic design and so these can be ignored. I am expecting to check and maybe adjust the Glyphs in punctuation in pink. In symbols accountof, addressedtothesubject, careof, need to be made to relate better but I think we are not kerning these.